+++ packages/react/src/alert-dialog/index.ts
@@ -1,0 +1,17 @@
+export {
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogContent,
+  AlertDialogOverlay,
+  AlertDialogCloseButton,
+} from "@chakra-ui/react";
+export type {
+  ModalContentProps,
+  ModalBodyProps,
+  ModalFooterProps,
+  ModalHeaderProps,
+  ModalOverlayProps,
+  CloseButtonProps,
+} from "@chakra-ui/react";