+++ packages/react/src/menu/index.ts
@@ -1,0 +1,20 @@
+export {
+  Menu,
+  MenuButton,
+  MenuList,
+  MenuItem,
+  MenuItemOption,
+  MenuGroup,
+  MenuOptionGroup,
+  MenuDivider,
+} from "@chakra-ui/react";
+export type {
+  MenuProps,
+  MenuButtonProps,
+  MenuListProps,
+  MenuItemProps,
+  MenuItemOptionProps,
+  MenuGroupProps,
+  MenuOptionGroupProps,
+  MenuDividerProps,
+} from "@chakra-ui/react";