+++ packages/react/src/typography/index.tsx
@@ -1,2 +1,2 @@
-export { Text, Heading } from "@chakra-ui/react";
-export type { TextProps, HeadingProps } from "@chakra-ui/react";
+export { Text, Heading, Highlight } from "@chakra-ui/react";
+export type { TextProps, HeadingProps, HighlightProps } from "@chakra-ui/react";